+++ server/routes/authRouter.js
@@ -1,0 +1,13 @@
+import express from 'express';
+import userController from '../controllers/authController';
+import inputValidator from '../middlewares/validator';
+
+const router = express.Router();
+
+// User sign up route
+router.post('/signup', inputValidator.userSignup, userController.signUp);
+
+// User sign in route
+router.post('/signin', inputValidator.userSignin, userController.signIn);
+
+export default router;
